Updated root route and tweaked css
