+++ frontend/src/components/NavBar/NavBar.css
@@ -10,3 +10,9 @@
     margin: 1em 1em;
 
 }
+
+.links {
+    display: flex;
+    gap: 0.5em;
+}
+
