+++ frontend/src/components/TvShowList/TvShowItem.css
@@ -4,7 +4,7 @@
     justify-content: space-around;
     border: 0.1em;
     border-style: solid;
-    margin-bottom: 1em;
+    margin: 1em 0em;
     border-radius: 5px;
     padding: 2em;
 }
