+++ frontend/src/routes/LandingPage.css
@@ -1,0 +1,14 @@
+p {
+    font-size: 0.8em;
+    margin-bottom: 0.5em;
+}
+h1 {
+    margin-bottom: 1em;
+}
+
+.landing-page-button {
+    display: flex;
+    justify-content: center;
+    gap: 1em;
+    margin: 1em 0em;
+}
